create a dialog for making sure that user want to delete a selected list
